% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getting_data}
\alias{getting_data}
\title{@import tidyverse, rvest, here, openxlsx, ,scholar, stringi, xml2
 @title data_getting
 @param df A dataframe that has group's informations
 @details This function get information from GrupLac}
\usage{
getting_data(df)
}
\description{
@import tidyverse, rvest, here, openxlsx, ,scholar, stringi, xml2
 @title data_getting
 @param df A dataframe that has group's informations
 @details This function get information from GrupLac
}
